Moved logger handling into single module to import
