+++ examples/api/brunel.py
@@ -17,7 +17,7 @@ from pyNN.utility import SimulationProgressBar
 import sys
 argv = sys.argv[1:]
 from pype9.mpi import is_mpi_master  # @IgnorePep8
-import pype9.utils.print_logger  # @UnusedImport @IgnorePep8
+import pype9.utils.logger_handlers.sysout_info  # @UnusedImport @IgnorePep8
 
 parser = argparse.ArgumentParser()
 parser.add_argument('--case', type=str, default='AI',
