+++ examples/api/izhikevich.py
@@ -8,7 +8,7 @@ import ninemlcatalog
 import quantities as pq
 import numpy as np
 import neo
-import pype9.utils.print_logger  # @UnusedImport
+import pype9.utils.logger_handlers.sysout_info  # @UnusedImport
 
 min_delay = 0.1
 max_delay = 10.0
