+++ test/test_seeding.py
@@ -3,7 +3,6 @@ from __future__ import print_function
 from itertools import chain
 import ninemlcatalog
 import numpy
-import logging
 import sys
 from nineml import units as un, Property
 from pype9.simulate.neuron import (
@@ -13,18 +12,11 @@ from pype9.simulate.nest import (
     CellMetaClass as NESTCellMetaClass, Network as NESTNetwork,
     Simulation as NESTSimulation)
 from pype9.simulate.neuron import Simulation
+import pype9.utils.logger_handlers.sysout_info  # @UnusedImport
 if __name__ == '__main__':
     from pype9.utils.testing import DummyTestCase as TestCase  # @UnusedImport
 else:
     from unittest import TestCase  # @Reimport
-
-logger = logging.getLogger('pype9')
-logger.setLevel(logging.INFO)
-handler = logging.StreamHandler(sys.stdout)
-formatter = logging.Formatter(
-    '%(levelname)s - %(message)s')
-handler.setFormatter(formatter)
-logger.addHandler(handler)
 
 
 class TestSeeding(TestCase):
