+++ test/test_simulations/test_cells.py
@@ -23,19 +23,11 @@ from pype9.simulate.nest import (  # @IgnorePep8
     Simulation as NESTSimulation)
 from pype9.utils.testing import Comparer, input_step, input_freq  # @IgnorePep8
 from pype9.simulate.nest.units import UnitHandler as UnitHandlerNEST  # @IgnorePep8
+import pype9.utils.logger_handlers.sysout_info  # @IgnorePep8
 if __name__ == '__main__':
     from pype9.utils.testing import DummyTestCase as TestCase  # @UnusedImport
 else:
     from unittest import TestCase  # @Reimport
-
-
-logger = logging.getLogger('pype9')
-logger.setLevel(logging.INFO)
-handler = logging.StreamHandler(sys.stdout)
-formatter = logging.Formatter(
-    '%(levelname)s - %(message)s')
-handler.setFormatter(formatter)
-logger.addHandler(handler)
 
 
 cell_metaclasses = {'neuron': NeuronCellMetaClass,
