+++ test/test_simulations/test_connect.py
@@ -18,15 +18,7 @@ if __name__ == '__main__':
     from pype9.utils.testing import DummyTestCase as TestCase  # @UnusedImport
 else:
     from unittest import TestCase  # @Reimport
-
-
-logger = logging.getLogger('pype9')
-logger.setLevel(logging.INFO)
-handler = logging.StreamHandler(sys.stdout)
-formatter = logging.Formatter(
-    '%(levelname)s - %(message)s')
-handler.setFormatter(formatter)
-logger.addHandler(handler)
+import pype9.utils.logger_handlers.sysout_info  # @IgnorePep8
 
 
 class TestConnect(TestCase):
